+++ projects/_03_independent_cqc/_02_clean/fargate/utils/clean_ct_outliers/clean_ct_care_home_outliers.py
@@ -4,7 +4,7 @@ from projects._03_independent_cqc._02_clean.fargate.utils.clean_ct_outliers.clea
     clean_longitudinal_outliers,
 )
 from projects._03_independent_cqc._02_clean.fargate.utils.clean_ct_outliers.clean_ct_repetition import (
-    clean_ct_values_after_consecutive_repetition,
+    null_values_exceeding_repetition_limit,
 )
 from projects._03_independent_cqc._02_clean.fargate.utils.clean_ct_outliers.null_posts_per_bed_ratio_outliers import (
     null_posts_per_bed_outliers,
@@ -51,18 +51,16 @@ def clean_capacity_tracker_care_home_outliers(lf: pl.LazyFrame) -> pl.LazyFrame:
 
     lf = null_posts_per_bed_outliers(lf)
 
-    lf = clean_longitudinal_outliers(
+    lf = null_values_exceeding_repetition_limit(
         lf=lf,
-        col_to_clean=IndCQC.ct_care_home_total_employed_cleaned,
-        cleaned_column_name=IndCQC.ct_care_home_total_employed_cleaned,
-        proportion_to_filter=0.001,
+        column_to_clean=IndCQC.ct_care_home_total_employed_cleaned,
         care_home=True,
     )
 
-    lf = clean_ct_values_after_consecutive_repetition(
+    lf = clean_longitudinal_outliers(
         lf=lf,
         column_to_clean=IndCQC.ct_care_home_total_employed_cleaned,
-        cleaned_column_name=IndCQC.ct_care_home_total_employed_cleaned,
+        proportion_to_filter=0.001,
         care_home=True,
     )