+++ projects/_03_independent_cqc/_02_clean/tests/fargate/utils/clean_ct_outliers/test_clean_ct_care_home_outliers.py
@@ -1,6 +1,5 @@
 import unittest
 from unittest.mock import Mock, patch
-import polars as pl
 
 import projects._03_independent_cqc._02_clean.fargate.utils.clean_ct_outliers.clean_ct_care_home_outliers as job
 
@@ -13,8 +12,8 @@ class CleanCapacityTrackerCareHomeOutliersTests(unittest.TestCase):
     def setUp(self) -> None:
         self.test_lf = Mock(name="ind_cqc_df")
 
-    @patch(f"{PATCH_PATH}.clean_ct_values_after_consecutive_repetition")
     @patch(f"{PATCH_PATH}.clean_longitudinal_outliers")
+    @patch(f"{PATCH_PATH}.null_values_exceeding_repetition_limit")
     @patch(f"{PATCH_PATH}.null_posts_per_bed_outliers")
     @patch(f"{PATCH_PATH}.aggregate_values_to_provider_level")
     @patch(f"{PATCH_PATH}.add_filtering_rule_column")
@@ -24,12 +23,12 @@ class CleanCapacityTrackerCareHomeOutliersTests(unittest.TestCase):
         aggregate_values_to_provider_level_mock: Mock,
         null_posts_per_bed_outliers_mock: Mock,
         clean_longitudinal_outliers_mock: Mock,
-        clean_ct_values_after_consecutive_repetition_mock: Mock,
+        null_values_exceeding_repetition_limit_mock: Mock,
     ):
         job.clean_capacity_tracker_care_home_outliers(self.test_lf)
 
         add_filtering_rule_column_mock.assert_called_once()
         aggregate_values_to_provider_level_mock.assert_called_once()
         null_posts_per_bed_outliers_mock.assert_called_once()
+        null_values_exceeding_repetition_limit_mock.assert_called_once()
         clean_longitudinal_outliers_mock.assert_called_once()
-        clean_ct_values_after_consecutive_repetition_mock.assert_called_once()